another commit

+++ src/main/frontend/src/meetings/MeetingsPage.js
@@ -1,11 +1,11 @@
-import {useEffect, useState} from "react";
+import { useEffect, useState } from "react";
 import NewMeetingForm from "./NewMeetingForm";
 import MeetingsList from "./MeetingsList";
 
 export default function MeetingsPage({username}) {
     const [meetings, setMeetings] = useState([]);
+    const [participants, setParticipants] = useState([]);
     const [addingNewMeeting, setAddingNewMeeting] = useState(false);
-    const [addingParticipants, setAddingParticipants] = useState(false);
 
     async function handleNewMeeting(meeting) {
         const response = await fetch(`/api/meetings`, {
@@ -22,24 +22,62 @@ export default function MeetingsPage({username}) {
     }
 
     useEffect(() => {
-        const fetchMeetings = async () => {
-            const response = await fetch(`/api/meetings`);
-            if (response.ok) {
-                const meetings = await response.json();
-                setMeetings(meetings);
-            }
-        };
+
         fetchMeetings();
     }, []);
 
-    async function handleDeleteMeeting(meeting) {
-        const response = await fetch(`/api/meetings/${meeting.id}`, {
-            method: 'DELETE',
-        });
+    async function fetchMeetings() {
+        const response = await fetch(`/api/meetings`);
         if (response.ok) {
-            const nextMeetings = meetings.filter(m => m !== meeting);
-            setMeetings(nextMeetings);
+            const meetings = await response.json();
+            setMeetings(meetings);
         }
+    }
+
+    async function handleDeleteMeeting(meeting) {
+        const responseOfParticipants = await fetch(`/api/meetings/${meeting.id}/participants`,{
+            method: 'GET',
+
+        })
+        const participants = await responseOfParticipants.json();
+
+        if (participants.length > 0) {
+            console.log("Nie można usunąć spotkania – są zapisani uczestnicy.");
+        }else{
+            const response = await fetch(`/api/meetings/${meeting.id}`, {
+                method: 'DELETE',
+            });
+            if (response.ok) {
+                const nextMeetings = meetings.filter(m => m !== meeting);
+                setMeetings(nextMeetings);
+            }
+        }
+    }
+
+    async function handleAddParticipant(meeting) {
+        await fetch(`/api/participants/`,{
+            method: 'POST',
+            body: JSON.stringify({login: username}),
+            headers: { 'Content-Type': 'application/json' }
+        })
+        await fetch(`/api/meetings/${meeting.id}/participants`, {
+            method: 'POST',
+            body: JSON.stringify({login: username}),
+            headers: { 'Content-Type': 'application/json' }
+        });
+
+        await fetchMeetings();
+    }
+
+    async function handledDeleteParticipant(meeting){
+        await fetch(`/api/meetings/${meeting.id}/participants/${username}`, {
+            method: 'DELETE'
+        });
+        await fetch(`/api/participants/${username}`, {
+            method: 'DELETE'
+        });
+
+        await fetchMeetings();
     }
 
 
@@ -53,7 +91,9 @@ export default function MeetingsPage({username}) {
             }
             {meetings.length > 0 &&
                 <MeetingsList meetings={meetings} username={username}
-                              onDelete={handleDeleteMeeting}/>}
+                              onButtonClick={handleAddParticipant}
+                              onButtonDelete={handledDeleteParticipant}
+                              onDelete={handleDeleteMeeting} />}
         </div>
     )
 }
